requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1

# Base de datos
pymongo==4.6.0
python-dotenv==1.0.0

# NLP y Machine Learning
transformers==4.35.2
torch==2.1.1

# Visualización
streamlit==1.29.0
pandas==2.1.3
altair==5.1.2
matplotlib==3.8.2
wordcloud==1.9.3

# Utilidades
lxml==4.9.3